Made painter signature more modern, added special case for liner gradients

+++ src/libnr/nr-gradient.h
@@ -33,7 +33,7 @@ struct _NRLGradientRenderer {
 	NRRenderer renderer;
 	const unsigned char *vector;
 	unsigned int spread;
-	float x0, y0;
+	int x0, y0;
 	float dx, dy;
 };
